requirements.txt
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==2.9.2
dnspython==2.1.0
Faker==6.1.1
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Naked==0.1.31
pycparser==2.20
pycryptodome==3.10.1
pymongo==3.11.3
pymongocrypt==1.0.1
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
shellescape==3.8.1
six==1.15.0
text-unidecode==1.3
urllib3==1.26.3
uWSGI==2.0.19.1
Werkzeug==1.0.1
